Array problem solved
Finally solved the 'array' problem

+++ ELL/ELL/TestPassVariables.h
@@ -7,14 +7,14 @@ struct typTmp {
 	int b = 0;
 	double c = 0;
 	int l = 0;
-	int r = 0;			// I can't make this a double. Are C++ and VBA doubles not matching? --> need to use PTRSAFE
+	double r = 0;			
 	//int arr[5];
 	int e = 0;
 };
 #pragma pack() // restore default
 
 
-int __stdcall ellPassStruct(typTmp &tmp);
-int __stdcall ellPassStruct02(typTmp &tmp);
+double __stdcall ellPassStruct(typTmp &tmp);
+double __stdcall ellPassStruct02(typTmp &tmp);
 void __stdcall ellPassStruct03(typTmp &tmp); 
 void __stdcall ellChangeRemotely(typTmp &tmp);
